+++ C++ Library/PathfindingLib/PathfindingLib/PGraphNode.h
@@ -6,7 +6,7 @@ class PGraphNode
 public:
 	PGraphNode();
 	~PGraphNode();
-	blaze::StaticVector<float, 3UL> position;
+	blaze::StaticVector<double, 3UL> position;
 	boolean obstacle;
 private:
 	blaze::StaticVector<int, 3UL> index;
